Give DebuggerBreak an x64 path + finish operator new sweep (x64 fix)
DebuggerBreak() was `_asm { int 3 }`; guard it with VER_ASM and use the
__debugbreak() intrinsic on the non-asm (x64) path - functionally
identical (INT3 breakpoint) on both architectures. Used tree-wide, so
this unblocks every project. Also fix the last multi-line custom
operator new size param uint -> size_t in FrameHeap.hpp (missed by the
line-based grep in cd22ad7). Together these fully clear the AI project
on x64. Win32 default (VER_ASM=TRUE) build green; x64 (VER_ASM=FALSE)
diagnostic errors 736 -> 620, AI project 0 errors.

+++ jp2_pc/Source/Lib/EntityDBase/FrameHeap.hpp
@@ -156,7 +156,7 @@ public:
 //
 inline void* operator new
 (
-	uint  u_size_type,			// Amount of memory to allocate in bytes.
+	size_t  u_size_type,			// Amount of memory to allocate in bytes.
 	const CFrameHeap& frh_heap	// Frame heap to allocate from.
 )
 //
